+++ src/main/java/com/github/antiad/AntiAd/Adfinder.java
@@ -25,7 +25,11 @@ public class Adfinder {
     // ip pattern NEW PATTERN: http://regexr.com/396h5  OLD PATTERN: http://regexr.com?33l17
     private final Pattern ipPattern = Pattern.compile("(?:\\d{1,3}[.,-:;\\/()=?}+ ]{1,4}){3}\\d{1,3}");
     // web pattern http://regexr.com?36elv
-    private final Pattern webpattern = Pattern.compile("[-a-zA-Z0-9@:%_\\+.~#?&//=]{2,256}\\.[a-z]{2,4}\\b(\\/[-a-zA-Z0-9@:%_\\+~#?&//=]*)?");
+    private Pattern webpattern;
+    // web pattern http://regexr.com?36elv
+    private final String webpatternAdvanced = "[-a-zA-Z0-9@:%_\\+.~#?&//=]{2,256}\\.[a-z]{2,4}\\b(\\/[-a-zA-Z0-9@:%_\\+~#?&//=]*)?";
+    // Simple pattern http://regexr.com/3cu3l
+    private final String webpatternSimple = "[-a-zA-Z0-9@:%_\\+.~#?&//=]{2,256}\\.(com|ru|net|org|de|jp|uk|br|pl|in|it|fr|au|info|nl|cn|ir|es|cz|biz|ca|kr|eu|ua|za|co|gr|ro|se|tw|vn|mx|ch|tr|at|be|hu|dk|tv|me|ar|us|no|sk|fi|id|cl|nz|by|pt)\\b(\\/[-a-zA-Z0-9@:%_\\+.~#?&//=]*)?";
     private HashMap<Player, Integer> warn;
     public boolean urlDetection, spamDetection, IPDetection, checkWordLenght;
     public int numbers, procentCapital;
@@ -33,7 +37,6 @@ public class Adfinder {
 
     public Adfinder(AntiAd instance) {
         plugin = instance;
-        loadWhitelist();
         startUp();
     }
 
@@ -91,7 +94,7 @@ public class Adfinder {
                 }
             }
         }
-        check.setMessage(StringUtils.join(words," "));
+        check.setMessage(StringUtils.join(words, " "));
         return caps;
     }
 
@@ -490,5 +493,14 @@ public class Adfinder {
         procentCapital = plugin.getConfig().getInt("Spam-Procent-Capital-Words");
         checkWordLenght = plugin.getConfig().getBoolean("Spam-Number-Letters-check");
         warn = new HashMap<Player, Integer>();
+        if (plugin.getConfig().getBoolean("useSimpleWebPattern")) {
+            webpattern = Pattern.compile(webpatternSimple);
+            plugin.debug("Simple pattern loaded");
+        } else {
+            webpattern = Pattern.compile(webpatternAdvanced);
+            plugin.debug("Advanced pattern loaded");
+        }
+
+        loadWhitelist();
     }
 }