+++ src/main/java/com/github/antiad/AntiAd/AntiAd.java
@@ -283,4 +283,9 @@ public class AntiAd extends JavaPlugin {
             }
         }
     }    
+
+    void reload() {
+        this.reloadConfig();
+        this.adfinder.startUp();
+    }
 }