+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/), and following
 ## [Unreleased][unreleased]
 ### Added
  - moved my debug into a config option
+ - Added a simple URL detection.
  
 ### Changed
  - Cleanup in the languages setup